chatbot_backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
anthropic
asyncpg
pymongo
reportlab
matplotlib
fpdf2
